Add `SolidBuilder`
