+++ crates/fj-kernel/src/builder/mod.rs
@@ -3,9 +3,11 @@
 mod cycle;
 mod edge;
 mod face;
+mod solid;
 
 pub use self::{
     cycle::CycleBuilder,
     edge::EdgeBuilder,
     face::{FaceBuilder, FacePolygon},
+    solid::SolidBuilder,
 };
