+++ crates/fj-kernel/src/objects/solid.rs
@@ -1,10 +1,8 @@
 use std::collections::BTreeSet;
 
-use fj_math::Scalar;
+use crate::builder::SolidBuilder;
 
-use crate::algorithms::TransformObject;
-
-use super::{Face, Surface};
+use super::Face;
 
 /// A 3-dimensional shape
 ///
@@ -32,28 +30,9 @@ impl Solid {
         Self { faces }
     }
 
-    /// Create a cube from the length of its edges
-    pub fn cube_from_edge_length(edge_length: impl Into<Scalar>) -> Self {
-        // Let's define a short-hand for half the edge length. We're going to
-        // need it a lot.
-        let h = edge_length.into() / 2.;
-
-        let points = [[-h, -h], [h, -h], [h, h], [-h, h]];
-
-        const Z: Scalar = Scalar::ZERO;
-        let planes = [
-            Surface::xy_plane().translate([Z, Z, -h]), // bottom
-            Surface::xy_plane().translate([Z, Z, h]),  // top
-            Surface::xz_plane().translate([Z, -h, Z]), // front
-            Surface::xz_plane().translate([Z, h, Z]),  // back
-            Surface::yz_plane().translate([-h, Z, Z]), // left
-            Surface::yz_plane().translate([h, Z, Z]),  // right
-        ];
-
-        let faces =
-            planes.map(|plane| Face::build(plane).polygon_from_points(points));
-
-        Solid::from_faces(faces)
+    /// Build a solid using [`SolidBuilder`]
+    pub fn build() -> SolidBuilder {
+        SolidBuilder
     }
 
     /// Access the solid's faces
